feature(agent): introduce concept of 'remote user' for agent connections
In fact this has been a tech-debt since the beginning of the prototype:
on the destination side of a remote operation, always the admin user (id=1)
was used (for destination item permissions/restrictions checks & retrieving
resources (directory listings)).

The concept of 'remote user' means that the remote user credentials need
to be provided when registering a new agent connection which are then
verified by the remote. If successful, the remote user's ID and the
combined server root and user scope are returned and persisted with
the connection entry, and remote operations will use the referred user
on the destination side for the given connection.