feat(agent): add dynamic user fs root path handling
User scope may be restricted to a file system path within Filebrowser's
root scope (similar to chroot'ing that user). Both the source and
destination's respective user may have such restriction in place
so during a remote copy operation these must be used to correctly
insert paths in the archive file at the source and to extrac
the archive to the correct path at the destination.