feat(agent): move from agent secret to temporary access token
The concept of using an 'Agent secret' to authenticate the first
session (during which the key-exchage is to happen) was inherently
flawed because the Agent secret was used in a password based SSH
authentication whereby the password was a preset constant.

To almost completely eliminate the threat from leaking such secret
used for this first time authentication, a 'Temporary Access Token'
concept was brought in. This is essentially a 'managed' SSH key-pair
created on demand whose secret may be shared with the party wishing
to connect.

This key is immediatelly revoked following a successful key exchange
or automatically after a short timeout.