feat(agent): add auth token to upload status update requests
Since it is now required for the remote resource and remote copy
operations that the user's JWT is submitted with the request,
it can be added to those requests that the Agent backend sends
to the Files api to update upload statuses. This, in turn,
allows the request handler that processes these status updates
in the Files backend to require a valid user session (which
previously hasn't been the case).