feat(agent): Change connection registration to only require a token
As opposed to the previous approach, username and password are not
required at the time of new Agent Connection registration. Instead,
each token is associated with its owner (the Files user who created
it) and the token owner becomes the remote user of the connection.

In other words, the user who creates the connection cannot choose
the remote user but it's determined by the token.

The connection owner must enter the remote user's password before
first use or every time the remote session expires.