feat(agent): add a better 'cancel transfer' implementation
Since remote file transfers are executed as a series of several tasks,
some of which are run in the main thread, some in Rust's or Tokio's
'green threads' and some in separate OS threads, and because a cancel
request may come from the user in any of these task execution phases,
a mutable state is required that can be shared between threads.

Every transfer registers a flag for itself in this shared state that
another request can modify. The transfer then monitors this flag
throughout the various phases and interrupts the process accordingly
when its flag has been set.